requirements.txt
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
redis==5.0.1
uvloop==0.19.0